+++ Oracle.NoSQL.SDK/src/Result/IResult.cs
@@ -22,6 +22,8 @@ namespace Oracle.NoSQL.SDK
 
         RowVersion ExistingVersion { get; set; }
 
+        DateTime? ExistingCreationTime { get; set; }
+
         DateTime? ExistingModificationTime { get; set; }
 
         string ExistingLastWriteMetadata { get; set; }